+++ routers/yiweixin/tasks.js
@@ -1,0 +1,24 @@
+var express = require('express');
+var app = express.Router();
+var models  = require('../../models')
+var helpers = require("../../helpers")
+var async = require("async")
+var requireLogin = helpers.requireLogin
+
+app.get('/tasks', function(req, res) {
+  var tasks = models.FlowTask.scope('active', 'defaultSort').findAll().then(function(tasks) {
+    res.render('yiweixin/flowtasks/index', { tasks: tasks })
+  }).catch(function(err) {
+      console.log(err)
+      res.redirect('/500')
+  })
+})
+
+app.get('/tasks/:id', function(req, res) {
+  models.FlowTask.findById(req.params.id).then(function(flowtask) {
+    res.render('yiweixin/flowtasks/show', { task: flowtask})
+  })
+})
+
+
+module.exports = app;
